requirements.txt
## pip install -r requirements.txt

aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
av==10.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
configparser==5.3.0
discord==2.1.0
discord.py==2.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
Pillow==9.4.0
pycparser==2.21
python-dotenv==1.0.0
redis==4.5.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
yarl==1.8.2
